+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to Merlin Spellbook will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Central composite sampling in `make-samples`.
+
 ## [0.2.1]
 
 ### Fixed